+++ src/NetCoreMQTTExampleCluster.SiloHost/OrleansAdoNetContent/SQLServer/SQLServer-Reminders.sql
@@ -1,11 +1,11 @@
--- Orleans Reminders table - https://dotnet.github.io/orleans/Documentation/Core-Features/Timers-and-Reminders.html
+-- Orleans Reminders table - https://docs.microsoft.com/dotnet/orleans/grains/timers-and-reminders
 CREATE TABLE OrleansRemindersTable
 (
 	ServiceId NVARCHAR(150) NOT NULL,
 	GrainId VARCHAR(150) NOT NULL,
 	ReminderName NVARCHAR(150) NOT NULL,
 	StartTime DATETIME2(3) NOT NULL,
-	Period INT NOT NULL,
+	Period BIGINT NOT NULL,
 	GrainHash INT NOT NULL,
 	Version INT NOT NULL,
